+++ BeatTogether.DedicatedServer.Node/Abstractions/IInstanceFactory.cs
@@ -12,9 +12,9 @@ namespace BeatTogether.DedicatedServer.Node.Abstractions
             bool permanentManager = false,//If a user links there account to discord and uses a bot to make a lobby, then can enter there userId
             float instanceTimeout = 0f,
             string ServerName = "",
-            float resultScreenTime = 20.0f,
-            float BeatmapStartTime = 5.0f,
-            float PlayersReadyCountdownTime = 0f,
+            long resultScreenTime = 20,
+            long BeatmapStartTime = 5,
+            long PlayersReadyCountdownTime = 0,
             bool AllowPerPlayerModifiers = false,
             bool AllowPerPlayerDifficulties = false,
             bool AllowChroma = true,
